Add AbstractTaskManagerTest
